created small component for inside loop
